Make sure state_nlm_notify properly handles stale cache inodes and exports
Now that refcounting is improved, this is actually much easier.

In state_nlm_notify we don't actually worry about stale cache inodes since
state_unlock will properly detect a stale cache inode and we can quietly
ignore the error here.

We do skip calling state_unlock() if the export is being removed since
state_unlock() does not detect that.

In either case, we can just ignore the error. We don't even need to count it
for early termination since these errors will not result in never draining
the list of locks belonging to the nsm_client.

We no longer need to take a reference on the lock entry since we only
look at it while holding the ssc_mutex (which assures the lock entry can
not become invalid since to do so, the ssc_mutex would need to be held to
remove the entry from the list of locks for the nsm_client).

Interestingly, we also don't have to worry about the locks that belong to the
current NSM state that are temporarily not on the nsm_client list. If they
get removed due to some other lock operation, they will cleanly be removed from
the newlocks list (even though that list is a stack local).

There is one vulnerability if a new SM_NOTIFY comes in while we are still
processing the one at hand. The new call to state_nlm_notify would miss the
locks on the newlocks list (which suggests that list should be attached to
the nsm_client). There is another issue in that now some THIRD NSM state is
the current state. The first state_nlm_notify will remove locks belonging to
that third state, suggesting that we need a way to terminate a
state_nlm_notify.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>