Use proper refcounting of state_t, state_owner_t, and export in layoutrecall.
Take a reference on each state_t in the recall list.

Take a reference to the state_owner_t to assure that is good also
before using it.

Use get_state_export_ref to properly reference state_export and handle
an export going stale.

Also, since the recall list is for a single cache inode entry, we don't
actually need to look at state_entry.

Also make some cosmetic changes to make the code easier to follow.

Also, fix a memory free ordering bug (free the handle FIRST, then the
cb_data that holds the layout recall that holds the pointer to the handle).

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>