Client manager hostaddr_str should just be IP address, not port also
The port may not be correct if the client has connected from a different
port either for session trunking or because it dropped the connection and
has created a new connection, so there is no point in printing it out.

hostaddr_str is only ever used in log messages, so this change has no
functional impact.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>