Convert rest of SAL display functions to use display_buffer
Also assure that if a string is constructed inside an if (isDebug())
construct, that we keep track of the validity of the string so that if the
log level increases, we don't then print garbage because the string
construction was skipped earlier.

Also, most places where a pointer to a SAL type is printed with %p,
display the referenced object instead.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>