CMake: check for attr/xattr.h when FSAL_GLUSTER gets build
Commit 6a466b7b adds attr/xattr.h to the #include files in FSAL_GLUSTER.
If this header is missing, compiling FSAL_GLUSTER would fail.

This change adds a check to CMakeLists.txt so that running cmake will
catch the missing header early.

On RPM based systems the libattr-devel package provides the attr/xattr.h
header. The .spec has also been updated to have this package as a
BuildRequires.

Reported-by: Meghana Madhusudhan <mmadhusu@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>
Acked-by: Meghana Madhusudhan <mmadhusu@redhat.com>
Acked-by: Soumya Koduri <skoduri@redhat.com>
Acked-by: Anand Subramanian <anands@redhat.com>