Napalm - Rip out cache_inode
Compile out cache_inode, and replace it with a thin wrapper in
fsal_helper.c that provides the same semantics to the upper layers.

Use fsal_helper code for NFSv4.

Make NFSv3 and NLM optional and disable, since they have not been
converted to the new fsal_helper API.  9P also fails to build, but it
was already optional.

Change-Id: Ifbb128a932d464e833fa45d71176ce050245fb22
Signed-off-by: Daniel Gryniewicz <dang@cohortfs.com>