Fix recursive locks when cleaning state
State cleanup can result in a recursive lock.  This was handled before
by a flag passed to cache_inode_lru_unref().  Until state refcounting
can be implemented, add an no_cleanup flag to state_hdl that is
protected by the state_lock, and use it to indicate when cleanup should
not occur.

Change-Id: Id96181964250d0663add2f67e479d768aa4bb96c
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>