Patch 0002-FSAL_GLUSTER-Add-option-to-configure-upcall-poll-int.patch
Subject: [PATCH] FSAL_GLUSTER: Add option to configure upcall poll interval

Currently upcall thread polls every 10us which results in
consuming lot of CPU at times. To avoid that provide an option
to be able to configure the poll interval so that users can
set it to optimal value as per their workloads.

The default value is 10us and can be increased upto 60sec.