Patch 0014-Allow-NLM_SHARE-access-none-deny-something-deny-read.patch
Subject: [PATCH 14/17] Allow NLM_SHARE access none, deny something (deny read
 for example)

It turns out Windows issues an NLM_SHARE access none, deny read in
the process of deleting a file. In order to push this down to the
FSAL, we need to have SOME kind of open mode, so we open for read.

This prevents an ugly crash...