Subject: [PATCH] Disallow client to create dirents with an empty name
NFSv4 spec forbids this, and Linux doesn't support it on local
filesystems, and it's caused problems with Gluster (which returns the
parent directory for an empty dirent name), so just protect against it.

Also, NFSv4 spec requires the error to be NFS4ERR_INVAL, not
NFS4ERR_BADNAME, so fix that as well.

Change-Id: Iebcc589ad3f131fd2a4ea500110550741eeb6d07