Fix attribute comparison in NFS4_OP_VERIFY
Our attribute comparison was broken.  It was comparing supported_attrs
correctly, but otherwise it was only ever comparing the first attribute
over and over.

In addition, it was assuming that the client creates the bitmaps and
attributes exactly like Ganesha does.  However some clients, for
example VMWare, generate bitmaps that are longer than necessary and
zero-filled.  This is correct, and we should handle it.

The new code walks the two bitmaps in parallel, compares set bits, and
then compares the value for attributes with both bits set.

Change-Id: Ia72125b64f949a3f9cc3f4e2269f7e5edb44db4c
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit 770d8c7539ef96efbfd6109604a3cf465009be35)
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>