MDCACHE - Release refs on dirents when chunk not consumed
If we don't consume all of a newly loaded chunk, then the refs on the
entries in that chunk are left over, and aren't cleared until the
directory is invalidated.

In this case, clear the refs on the entries when we bail early from a
chunk.

Change-Id: I61c6608898cb999c8440929504cf122b4bef232b
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit 136df4f262c3f9bc29df456eac50921321826da3)