MDCACHE - Fix chunk refcounting in readdir
Chunk refcounting was a bit loose, and there was a case where a chunk
could get reaped while in use, when the write lock was held.

Clean up the chunk refcounting, so that all the pointers to chunks that
we keep during readdir have a refcount, and that we drop all the refs
appropriately.

Change-Id: Ic3acfc71d15cde1cdf68a46864f5e60445c4ff71
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>

Merge

Change-Id: Ie76e1a6d2a31a420b037b37befec126901dc6997
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>